Move monitorLocalStatus to BaseWorkUnit
